requirements.txt
alembic==1.12.0
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
pytz==2023.3.post1
SQLAlchemy==2.0.20
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.9.0
